pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ouro-agents"
version = "0.1.0"
description = "Autonomous Agent Package for Ouro"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uuid7-standard>=1.1.0",
    "smolagents[mcp]",
    "rich",
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "apscheduler",
    "mem0ai",
    # Aligned with mem0 vector_stores extra; do not cap <1.0 — that downgrades vs existing on-disk data and breaks sysdb (KeyError: '_type').
    "chromadb>=0.4.24",
    "anthropic",
    "openai",
    "mcp",
    "textual",
    "ouro-mcp>=0.3.0",
    "ouro-py",
]

[project.scripts]
ouro-agents = "ouro_agents.runner:main"

[project.optional-dependencies]
dev = ["pytest", "ruff"]
